fix: resolve critical build errors after merge
- Fix effectiveDecoration parameter name in advanced_theme_showcase.dart
- Add missing dart:math import for max() function in custom_chat_widget.dart
- Fix type conversion from num to double for maxWidth
- Remove non-existent assets directory from pubspec.yaml

✅ Build now compiles successfully

+++ lib/src/widgets/custom_chat_widget.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -391,7 +392,7 @@ class _CustomChatWidgetState extends State<CustomChatWidget> {
     // Use different widths for user vs AI messages
     final maxWidth = isUser ? bubbleStyle.userBubbleMaxWidth ??
             (textSize.width < defaultMaxWidth 
-                ? (115 + max(textSize.width,max(usernameTextSize.width,timeTextSize.width)))
+                ? (115 + max(textSize.width,max(usernameTextSize.width,timeTextSize.width))).toDouble()
                 : defaultMaxWidth)
         : bubbleStyle.aiBubbleMaxWidth ??
             MediaQuery.of(context).size.width * 0.88;
